+++ src/AuthorizationMiddlewareFactory.php
@@ -10,12 +10,10 @@ declare(strict_types=1);
 namespace Zend\Expressive\Authorization;
 
 use Psr\Container\ContainerInterface;
-use Zend\Expressive\Authentication\ResponsePrototypeTrait;
+use Psr\Http\Message\ResponseInterface;
 
 class AuthorizationMiddlewareFactory
 {
-    use ResponsePrototypeTrait;
-
     public function __invoke(ContainerInterface $container) : AuthorizationMiddleware
     {
         if (! $container->has(AuthorizationInterface::class)) {
@@ -26,15 +24,9 @@ class AuthorizationMiddlewareFactory
             ));
         }
 
-        try {
-            $responsePrototype = $this->getResponsePrototype($container);
-        } catch (\Exception $e) {
-            throw new Exception\InvalidConfigException($e->getMessage());
-        }
-
         return new AuthorizationMiddleware(
             $container->get(AuthorizationInterface::class),
-            $responsePrototype
+            $container->get(ResponseInterface::class)
         );
     }
 }