+++ src/ConfigProvider.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-authorization for the canonical source repository
- * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (https://www.zend.com)
+ * @copyright Copyright (c) 2017-2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-authorization/blob/master/LICENSE.md New BSD License
  */
 
@@ -28,7 +28,7 @@ class ConfigProvider
     public function getDependencies() : array
     {
         return [
-            'factories'  => [
+            'factories' => [
                 AuthorizationMiddleware::class => AuthorizationMiddlewareFactory::class
             ],
         ];