+++ test/AuthorizationMiddlewareTest.php
@@ -11,6 +11,7 @@ namespace ZendTest\Expressive\Authorization;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
+use Prophecy\Prophecy\ObjectProphecy;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\RequestHandlerInterface;
@@ -23,37 +24,55 @@ class AuthorizationMiddlewareTest extends TestCase
 {
     use UserTrait;
 
+    /** @var AuthorizationInterface|ObjectProphecy */
+    private $authorization;
+
+    /** @var ServerRequestInterface|ObjectProphecy */
+    private $request;
+
+    /** @var RequestHandlerInterface|ObjectProphecy */
+    private $handler;
+
+    /** @var ResponseInterface|ObjectProphecy */
+    private $responsePrototype;
+
+    /** @var callable */
+    private $responseFactory;
+
     protected function setUp()
     {
         $this->authorization = $this->prophesize(AuthorizationInterface::class);
         $this->request = $this->prophesize(ServerRequestInterface::class);
         $this->handler = $this->prophesize(RequestHandlerInterface::class);
-        $this->response = $this->prophesize(ResponseInterface::class);
+        $this->responsePrototype = $this->prophesize(ResponseInterface::class);
+        $this->responseFactory = function () {
+            return $this->responsePrototype->reveal();
+        };
     }
 
     public function testConstructor()
     {
-        $middleware = new AuthorizationMiddleware($this->authorization->reveal(), $this->response->reveal());
+        $middleware = new AuthorizationMiddleware($this->authorization->reveal(), $this->responseFactory);
         $this->assertInstanceOf(AuthorizationMiddleware::class, $middleware);
     }
 
     public function testProcessWithoutUserAttribute()
     {
         $this->request->getAttribute(UserInterface::class, false)->willReturn(false);
-        $this->response->withStatus(401)->will([$this->response, 'reveal']);
+        $this->responsePrototype->withStatus(401)->will([$this->responsePrototype, 'reveal']);
 
         $this->handler
             ->handle(Argument::any())
             ->shouldNotBeCalled();
 
-        $middleware = new AuthorizationMiddleware($this->authorization->reveal(), $this->response->reveal());
+        $middleware = new AuthorizationMiddleware($this->authorization->reveal(), $this->responseFactory);
 
         $response = $middleware->process(
             $this->request->reveal(),
             $this->handler->reveal()
         );
 
-        $this->assertSame($this->response->reveal(), $response);
+        $this->assertSame($this->responsePrototype->reveal(), $response);
     }
 
     public function testProcessRoleNotGranted()
@@ -61,9 +80,9 @@ class AuthorizationMiddlewareTest extends TestCase
         $this->request
             ->getAttribute(UserInterface::class, false)
             ->willReturn($this->generateUser('foo', ['bar']));
-        $this->response
+        $this->responsePrototype
             ->withStatus(403)
-            ->will([$this->response, 'reveal']);
+            ->will([$this->responsePrototype, 'reveal']);
         $this->authorization
             ->isGranted('bar', Argument::that([$this->request, 'reveal']))
             ->willReturn(false);
@@ -72,14 +91,14 @@ class AuthorizationMiddlewareTest extends TestCase
             ->handle(Argument::any())
             ->shouldNotBeCalled();
 
-        $middleware = new AuthorizationMiddleware($this->authorization->reveal(), $this->response->reveal());
+        $middleware = new AuthorizationMiddleware($this->authorization->reveal(), $this->responseFactory);
 
         $response = $middleware->process(
             $this->request->reveal(),
             $this->handler->reveal()
         );
 
-        $this->assertSame($this->response->reveal(), $response);
+        $this->assertSame($this->responsePrototype->reveal(), $response);
     }
 
     public function testProcessRoleGranted()
@@ -93,15 +112,15 @@ class AuthorizationMiddlewareTest extends TestCase
 
         $this->handler
             ->handle(Argument::any())
-            ->will([$this->response, 'reveal']);
+            ->will([$this->responsePrototype, 'reveal']);
 
-        $middleware = new AuthorizationMiddleware($this->authorization->reveal(), $this->response->reveal());
+        $middleware = new AuthorizationMiddleware($this->authorization->reveal(), $this->responseFactory);
 
         $response = $middleware->process(
             $this->request->reveal(),
             $this->handler->reveal()
         );
 
-        $this->assertSame($this->response->reveal(), $response);
+        $this->assertSame($this->responsePrototype->reveal(), $response);
     }
 }